handle cases where the symlink target exists gracefully
If the link target already exists and points to where it's supposed
to, just print a warning.  It probably was listed both in
dependencies and devDependencies.

If the entry in node_modules is wrong, continue to fail hard, but
with a nice, informative error now.

closes #1 on GitHub

(cherry picked from commit 3af60603b460c93f157ca7adb5373b669f22dcba)