nodejs-symlink-deps: add --force option
if a bundled dep is encoutered, this option will force nodejs-symlink-deps
to delete the bundled dep and replace it with a symlink.

This option is useful for development, where you might have some deps already present
in the distribution but want to use bundled or npm installed copies for others.

This option should NEVER be used in offical Fedora spec files.

(cherry picked from commit dd772ada44fddcb1ef69483f9af467a29d749505)